feat: replace studio page
