+++ scripts/projects/bodies/epic-migration.md
@@ -1,17 +1,20 @@
 目标
+
 - 从模拟页迁移能力下沉为组件，提供节点级→字段级的迁移向导与护栏。
 
 范围
+
 - 迁移向导组件化与状态抽离；
 - 自动映射置信度标签/原因；
 - 撤销/审计写入事件流。
 
 验收
+
 - [ ] 可从 Inspector 触发向导，完成字段级映射并应用；
 - [ ] 事件流记录每步变更，可逐步回滚；
 - [ ] 自动映射命中率指标可观测。
 
 参考
-- `AGENTS.md` 迁移向导
-- `docs/工作流编排_studio（模拟页面）.jsx`
 
+- `AGENTS.md` 迁移向导
+- `src/studio/StudioPage.tsx`
